% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_importers.R
\name{ABS_read}
\alias{ABS_read}
\title{Read in ABS files.}
\usage{
ABS_read(
  ABS_path,
  order = TRUE,
  recursive = FALSE,
  dec = NULL,
  sep = NULL,
  verbose = TRUE,
  cores = parallel::detectCores(logical = FALSE)
)
}
\arguments{
\item{ABS_path}{full path to either (1) a folder containing multiple ABS .dat files, or (2) a single absorbance file.}

\item{order}{TRUE/FALSE to order data according to wavelength}

\item{recursive}{TRUE/FALSE recurse into sub-directories?}

\item{dec}{see ?staRdom::absorbance_read(). Table decimal separator testing.}

\item{sep}{see ?staRdom::absorbance_read(). Optional or automatic separator determination.}

\item{verbose}{TRUE/FALSE to provide additional information whilst running.}

\item{cores}{number of CPU cores used for parallel computation. Defaults to maximum, detected automatically.}
}
\description{
Read in one or more Absorbance (ABS) ASCII .dat files from a given directory. This function is a direct
   modification of absorbance_read() from the staRdom package, with adjustments made to the filename
   fetching lines (get_names instead of list.files) to ensure .dat files are correctly fetched.
}
