% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_functions.R
\name{create_queue_folders}
\alias{create_queue_folders}
\title{Create the sample queue folder framework.}
\usage{
create_queue_folders(parent_directory, type = "all")
}
\arguments{
\item{parent_directory}{the folder within which all the folders will be created}

\item{type}{one of "all", "import" or "export". "All" will create all the folders, and the other two options will only fill out those respective folders.}
}
\description{
Uses default_folder_layout() to create a set of nested folders that are used by the file handling functions. It does not overwrite; you can "fill out" an incomplete folder set.
}
