% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_file_functions.R
\name{generate_logfile}
\alias{generate_logfile}
\title{Save a log file as a .txt using write.table.}
\usage{
generate_logfile(run_sheet, destination, folder)
}
\arguments{
\item{run_sheet}{The run sheet for this folder.}

\item{destination}{A full path to the export directory.}

\item{folder}{}
}
\description{
Write a log file to the specified destination.
}
