% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_names}
\alias{get_names}
\title{Get object (file or folder) names from a given directory}
\usage{
get_names(
  directory,
  type = "files",
  string = NULL,
  full_names = TRUE,
  recursive = FALSE
)
}
\arguments{
\item{directory}{full path to the folder}

\item{type}{character string string, either "files" or "folders".}

\item{string}{either NULL or an optional character string to index the files or folders.}

\item{full_names}{TRUE/FALSE to obtain full paths for each object.}

\item{recursive}{list.dirs and list.files option. Should the listing recurse into directories?}
}
\description{
A wrapper for list.dirs and list.files. Obtains names of objects within the
      specified directory.
}
