% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_importers.R
\name{PCT_read}
\alias{PCT_read}
\title{Read in PCT files.}
\usage{
PCT_read(
  PCT_path,
  order = TRUE,
  recursive = FALSE,
  dec = NULL,
  str_match = TRUE,
  sep = NULL,
  verbose = TRUE,
  cores = parallel::detectCores(logical = FALSE)
)
}
\arguments{
\item{order}{TRUE/FALSE to order data according to wavelength.}

\item{recursive}{TRUE/FALSE to recurse into sub-directories.}

\item{dec}{see ?staRdom::absorbance_read(). Table decimal separator testing.}

\item{str_match}{Use simple string-matching to ensure the indexed absorbance files contain the string "PCT", as standard in Percent Transmission files from the Aqualog Sample Queue system.}

\item{sep}{see ?staRdom::absorbance_read(). Optional or automatic separator determination.}

\item{verbose}{TRUE/FALSE to provide additional information whilst running.}

\item{cores}{number of CPU cores used for parallel computation. Defaults to maximum, detected automatically.}

\item{ABS_path}{full path to either (1) a folder containing multiple PCT .dat files, or (2) a single absorbance file.}
}
\description{
Read in one or more Percent Transmission (PCT) ASCII .dat files from a given directory. This function is a direct
   modification of absorbance_read() from the staRdom package, with adjustments made to the filename
   fetching lines (get_names instead of list.files) to ensure .dat files are correctly fetched.
}
