% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sample_queue.R
\name{process_sample_queue}
\alias{process_sample_queue}
\title{Rename and collate Aqualog sample queue output files.}
\usage{
process_sample_queue(
  folder,
  export_dir,
  run_date,
  run_sheet,
  write_over = TRUE,
  mqblank_sub = TRUE,
  dilution_process = TRUE,
  neg_to_0 = TRUE,
  dry_run = FALSE
)
}
\arguments{
\item{folder}{The folder within the import directory containing all the files from the sample queue run related to the supplied run sheet.}

\item{export_dir}{The export directory containing the 'type' subfolders.}

\item{run_date}{Date of format dd/mm/yy on which the run was completed on the Aqualog.}

\item{run_sheet}{A data.frame matching the format of a run sheet. For info on how to format a run sheet, see the package documentation.}

\item{write_over}{TRUE/FALSE file.copy parameter. Write over identically named files in the destination folders?}

\item{mqblank_sub}{TRUE/FALSE. Imports sample types and subtracts an average of all milli-q blanks in the run from each. Exports .csv outputs to the "blank subtracted PEM" folder in each sample type folder.}

\item{dilution_process}{TRUE/FALSE where dilution factor values are given in the run sheet, perform a corresponding multiplication to the fluorescence intensity values of those EEMs. Defaults to TRUE.}

\item{neg_to_0}{TRUE/FALSE to set all instances of negative fluorescence in EEMs to 0 prior to export. Defaults to TRUE.}

\item{dry_run}{TRUE/FALSE to skip all file copying and return a log of all file import and export paths. Simulates a 'real' run.}
}
\description{
Copies, renames and sorts output files from the Aqualog sample queueing system based upon a supplied run sheet. Also creates a log file using generate_logfile() and copies, renames .opj project files using transfer_project_files(). Options for simple post-processing (e.g. blank subtraction)
}
