% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_file_functions.R
\name{transfer_project_files}
\alias{transfer_project_files}
\title{Transfer project files (.opj extension) to the project file folder, with some renaming.}
\usage{
transfer_project_files(foldername, run_date)
}
\arguments{
\item{foldername}{Name of the folder containing the project file and other target samples.}

\item{run_date}{The date of the run in format %d%m%y.}
}
\description{
Takes all the project files in the target folder and moves them to the project file folder inside the export directory.
}
