% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sample_queue.R
\name{process_sample_queue}
\alias{process_sample_queue}
\title{Rename and collate Aqualog sample queue output files.}
\usage{
process_sample_queue(
  folder,
  export_dir,
  run_date,
  run_sheet,
  write_over = TRUE
)
}
\arguments{
\item{folder}{The folder within the import directory containing all the files from the sample queue run related to the supplied run sheet.}

\item{export_dir}{The export directory containing the 'type' subfolders.}

\item{run_date}{Date of format dd/mm/yy on which the run was completed on the Aqualog.}

\item{run_sheet}{A data.frame matching the format of a run sheet. For info on how to format a run sheet, see the package documentation.}

\item{write_over}{TRUE/FALSE file.copy parameter. Write over identically named files in the destination folders?}
}
\description{
Copies, renames and sorts output files from the Aqualog sample queueing system based upon a supplied run sheet. Also creates a log file using generate_logfile() and copies, renames .opj project files using transfer_project_files().
}
